+++ c++App/Plotroc/rocdata.h
@@ -14,7 +14,7 @@ public:
     QVector<  double> getTprs(int pos) const;
     QVector<  double> getFprs(int pos) const;
     QList<QString> getPVFilenames() const;
-    QList<int> getFrequencies() const;
+    QList<double> getFrequencies() const;
 
 
     QVector<bool> getIsTruePositives() const;
@@ -34,7 +34,7 @@ private:
     QList<QVector< double>> pvVectors;
     QList<QVector< double>> rocTprs;
     QList<QVector< double>> rocFprs;
-    QList<int> frequencies;
+    QList<double> frequencies;
 
     QCommandLineParser& clp;
     //FrequencyPartition freqPartition;
